abstract ReadWriteLock into a reusable module
